// src/types/styled.ts
export interface StyledTextProps {
  $color?: string;
  $type: string;
  $weight?: string;
  $align?: string;
  $ellipsis?: boolean | number;
  $bullet?: boolean;
  $styled?: any;
  $whiteSpace?: string;
}

export interface StyledButtonProps {
  $type: string;
  $size: string;
  $fullWidth?: boolean;
}